requirements.txt
Authlib==1.6.0
fastapi[standard]==0.115.14
plotly==6.2.0
python-dotenv==1.1.1
streamlit==1.46.1